+++ src/InvestLens.Model/Stat.cs
@@ -2,7 +2,7 @@
 
 public class Stat
 {
-    public Stat(string title, double value, string unit="", bool unitIsSuffix=true)
+    public Stat(string title, decimal value, string unit="", bool unitIsSuffix=true)
     {
         Title = title;
         Value = value;
@@ -11,7 +11,7 @@ public class Stat
     }
 
     public string Title { get; set; }
-    public double Value { get; set; }
+    public decimal Value { get; set; }
     public string Unit { get; set; }
     public bool UnitIsSuffix { get; set; }
 }
